requirements.txt
runpod>=1.6.0
websocket-client>=1.6.0
boto3>=1.28.0
requests>=2.31.0
Pillow>=10.0.0
aiohttp>=3.9.0
